+++ src/errors/forbidden-json-api-error.ts
@@ -1,4 +1,5 @@
 import { JsonApiError } from '../json-api-error';
+import { DocumentObject } from '../types/document-object';
 
 /**
  * Forbidden json api error.
@@ -11,10 +12,11 @@ export class ForbiddenJsonApiError extends JsonApiError
      * Constructor.
      * 
      * @param {string} href Target href.
+     * @param {DocumentObject} documentObject Document object.
      */
-    public constructor(href: string)
+    public constructor(href: string, documentObject: DocumentObject)
     {
-        super(`Unsupported request ${href} is sent to create a resource.`, href, 403);
+        super(`Unsupported request ${href} is sent to create a resource.`, href, 403, documentObject);
         
         Object.setPrototypeOf(this, new.target.prototype);