+++ src/errors/not-found-json-api-error.ts
@@ -1,4 +1,5 @@
 import { JsonApiError } from '../json-api-error';
+import { DocumentObject } from '../types/document-object';
 
 /**
  * Not found json api error.
@@ -11,10 +12,11 @@ export class NotFoundJsonApiError extends JsonApiError
      * Constructor.
      * 
      * @param {string} href Target href.
+     * @param {DocumentObject} documentObject Document object.
      */
-    public constructor(href: string)
+    public constructor(href: string, documentObject: DocumentObject)
     {
-        super(`Target resource ${href} does not exists.`, href, 404);
+        super(`Target resource ${href} does not exists.`, href, 404, documentObject);
 
         Object.setPrototypeOf(this, new.target.prototype);