+++ src/errors/other-json-api-error.ts
@@ -1,4 +1,5 @@
 import { JsonApiError } from '../json-api-error';
+import { DocumentObject } from '../types/document-object';
 
 /**
  * Other json api error.
@@ -12,10 +13,11 @@ export class OtherJsonApiError extends JsonApiError
      * 
      * @param {string} href Target href.
      * @param {number} status Status.
+     * @param {DocumentObject} documentObject Document object.
      */
-    public constructor(href: string, status: number)
+    public constructor(href: string, status: number, documentObject: DocumentObject)
     {
-        super(`Error occured while accesing the resource ${href}.`, href, status);
+        super(`Error occured while accesing the resource ${href}.`, href, status, documentObject);
 
         Object.setPrototypeOf(this, new.target.prototype);