+++ src/json-api-error.ts
@@ -1,4 +1,5 @@
 import { EntityStoreError } from '@dipscope/entity-store';
+import { DocumentObject } from './types/document-object';
 
 /**
  * Represents json api error.
@@ -22,13 +23,21 @@ export class JsonApiError extends EntityStoreError
     public readonly status: number;
 
     /**
+     * Document object returned from api.
+     * 
+     * @type {DocumentObject}
+     */
+    public readonly documentObject: DocumentObject
+
+    /**
      * Constructor.
      * 
      * @param {string} message Message.
      * @param {string} href Href.
      * @param {number} status Status.
+     * @param {DocumentObject} status Document object.
      */
-    public constructor(message: string, href: string, status: number)
+    public constructor(message: string, href: string, status: number, documentObject: DocumentObject)
     {
         super(message);
 
@@ -36,6 +45,7 @@ export class JsonApiError extends EntityStoreError
 
         this.href = href;
         this.status = status;
+        this.documentObject = documentObject;
 
         return;
     }