+++ src/json-api-resource-metadata-key.ts
@@ -1,6 +1,9 @@
+import { CustomKey } from '@dipscope/type-manager';
+import { JsonApiResourceMetadata } from './json-api-resource-metadata';
+
 /**
  * Key used to store json api resource metadata inside type manager custom data.
  * 
- * @type {string}
+ * @type {CustomKey<JsonApiResourceMetadata<any>>}
  */
-export const jsonApiResourceMetadataKey: string = '$ESJsonApiResourceMetadata';
+export const jsonApiResourceMetadataKey: CustomKey<JsonApiResourceMetadata<any>> = new CustomKey('$DSJsonApiResourceMetadata');