+++ webpack.config.js
@@ -9,9 +9,9 @@ module.exports = [{
         index: './src/index.ts'
     },
     output: {
-        path: Path.resolve(__dirname, 'dist'),
+        path: Path.resolve(__dirname, 'dist/umd'),
         filename: '[name].js',
-        library: 'JsonApiEntityProvider',
+        library: '$DSJsonApiEntityProvider',
         libraryTarget: 'umd',
         globalObject: 'this',
         umdNamedDefine: true
@@ -27,13 +27,13 @@ module.exports = [{
             commonjs: '@dipscope/type-manager',
             commonjs2: '@dipscope/type-manager',
             amd: '@dipscope/type-manager',
-            root: 'TypeManager'
+            root: '$DSTypeManager'
         },
         '@dipscope/entity-store': {
             commonjs: '@dipscope/entity-store',
             commonjs2: '@dipscope/entity-store',
             amd: '@dipscope/entity-store',
-            root: 'EntityStore'
+            root: '$DSEntityStore'
         }
     },
     plugins: [
@@ -51,7 +51,7 @@ module.exports = [{
             use: [{
                 loader: 'ts-loader',
                 options: {
-                    configFile: 'tsconfig.webpack.json'
+                    configFile: 'tsconfig/tsconfig.umd.json'
                 }
             }],
             exclude: /node_modules/,
@@ -71,9 +71,9 @@ module.exports = [{
         index: './src/index.ts'
     },
     output: {
-        path: Path.resolve(__dirname, 'dist'),
+        path: Path.resolve(__dirname, 'dist/umd'),
         filename: '[name].browser.js',
-        library: 'JsonApiEntityProvider',
+        library: '$DSJsonApiEntityProvider',
         libraryTarget: 'umd',
         globalObject: 'this',
         umdNamedDefine: true
@@ -89,13 +89,13 @@ module.exports = [{
             commonjs: '@dipscope/type-manager',
             commonjs2: '@dipscope/type-manager',
             amd: '@dipscope/type-manager',
-            root: 'TypeManager'
+            root: '$DSTypeManager'
         },
         '@dipscope/entity-store': {
             commonjs: '@dipscope/entity-store',
             commonjs2: '@dipscope/entity-store',
             amd: '@dipscope/entity-store',
-            root: 'EntityStore'
+            root: '$DSEntityStore'
         }
     },
     plugins: [
@@ -113,7 +113,7 @@ module.exports = [{
             use: [{
                 loader: 'ts-loader',
                 options: {
-                    configFile: 'tsconfig.webpack.json'
+                    configFile: 'tsconfig/tsconfig.umd.json'
                 }
             }],
             exclude: /node_modules/,