+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2023-??-??
+
+### Added
+
+- Fluent api for declarative configuration with type manager.
+
+## [1.1.0] - 2023-03-14
+
+### Added
+
+- Support for custom resource routes.
+
 ## [1.0.1] - 2022-08-14
 
 ### Fixed